Generalize pal and ghost to players vector (#15)
* Refactor move

* Fix animations, store player vector

* Generalize pal and ghost to players vector

* Add player type

+++ source/GhostedApp.h
@@ -54,7 +54,7 @@ protected:
     StartScene _start;
     GameScene _gameplay;
 
-    unsigned _mode;
+    uint8_t _mode;
     
     /** 
      * Internal helper to build the scene graph.
